Fix errors for missing class_data_item
For some reason, some class analysis objects don't have class_data_item,
which causes problems since we are using it for field data retrieval.
It seems that for all of those classes there are no fields, so in this
commit we fix the errors by assuming it's true.
This assumption should be reviewed with more apks in the future!